Add keycloak config with env vars
