+++ frontend/.env-example
@@ -1,1 +1,3 @@
-API_URL=<insert api url>
+API_URL=http://localhost:5000
+KEYCLOAK_RESOURCE: 'KEYCLOAK-RESOURCE',
+KEYCLOAK_CLIENT_ID: 'KEYCLOAK-CLIENT'
