+++ frontend/server.js
@@ -45,6 +45,8 @@ app.get("/env", function(req, res) {
   res.send({
       backend: 'mds-python-backend',
       apiUrl: process.env.API_URL,
+      keycloak_resource: process.env.KEYCLOAK_RESOURCE,
+      keycloak_clientId: process.env.KEYCLOAK_CLIENT_ID
   });
 });
