+++ openshift/_nodejs.dc.json
@@ -41,6 +41,18 @@
       "value": "DEV"
     },
     {
+      "name": "KEYCLOAK_RESOURCE",
+      "displayName": "Resource ID",
+      "description": "Resource ID for keycloak",
+      "value": "mines-application-dev"
+    },
+    {
+      "name": "KEYCLOAK_CLIENT_ID",
+      "displayName": "Client ID",
+      "description": "Client ID for keycloak",
+      "value": "mines-application-dev"
+    },
+    {
       "name": "API_URL",
       "displayName": "API Url",
       "description": "URL for the backend API",
@@ -164,6 +176,14 @@
                   {
                     "name": "NODE_ENV",
                     "value": "${NODE_ENV}"
+                  },
+                  {
+                    "name": "KEYCLOAK_RESOURCE",
+                    "value": "${KEYCLOAK_RESOURCE}"
+                  },
+                  {
+                    "name": "KEYCLOAK_CLIENT_ID",
+                    "value": "${KEYCLOAK_CLIENT_ID}"
                   }
                 ],
                 "ports": [
